Remove the bootstrap process
To make it much easier for chef-load to
begin creating API load we are going to
just use the admin user to make all API
requests to the Chef Server. This means
we no longer need a bootstrap process
that creates client objects for every
node. The only downside I'm aware of is
that the Chef Server logs will show that
the admin user is making all of the API
requests.