Refactor directory structure to add sub-commands
This commit is refactoring a bunch of files, it is creating two main
directories, `commands/` and `lib/`.

The `commands/` directory will host all the commands that chef-load
will provide.

The `lib/` directory is meant to be the libraries that chef-load will
use to do specific tasks like generate nodes and start the continuous
load of nodes, actions, reports, etc.

The `lib/` directory will be able to be imported as an external library
to provide load functionality to other go applications.

Signed-off-by: Salim Afiune <afiune@chef.io>